code/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
matplotlib==3.9.2
Pillow==10.4.0
PyYAML==6.0.2
pyserial==3.5
smbus2==0.5.0
spidev==3.6
RPi.GPIO==0.7.1
gpiozero==1.7.2
pigpio==1.78
adafruit-circuitpython-servokit==1.3.16
adafruit-circuitpython-bno08x==1.2.11
adafruit-circuitpython-pca9685==3.4.10
ydlidar==1.0.2
rich==13.7.1
